requirements.txt
django == 4.2.8
django-cors-headers == 4.3.1
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.3.1
pymysql == 1.1.0
Pillow == 10.1.0
